+++ data-raw/DATASET.R
@@ -80,7 +80,7 @@ planets <- planets %>% dplyr::as_tibble() %>%
 usethis::use_data(planets, overwrite = T)
 
 ## Sample energy futures datasets
-df_fut <- readRDS("df_fut") ; usethis::use_data(df_fut, overwrite = T)
+#df_fut <- readRDS("df_fut") ; usethis::use_data(df_fut, overwrite = T)
 
 iuser = mstar[["iuser"]] ; ipassword = mstar[["ipassword"]]
 startdate <- "2004-01-01"
@@ -148,24 +148,23 @@ usethis::use_data(eiaStorageCap, overwrite = T)
 ## Sample GIS Mapping
 library(rgdal)
 
-crudepipelines <- rgdal::readOGR(dsn = "~/now/RTL/GIS_EIA/CrudeOil_Pipelines_US_EIA/", layer = "CrudeOil_Pipelines_US_202001")
-refineries <- rgdal::readOGR(dsn = "~/now/RTL/GIS_EIA/Petroleum_Refineries_US_EIA/", layer = "Petroleum_Refineries_US_2020")
-productspipelines <- rgdal::readOGR(dsn = "~/now/RTL/GIS_EIA/PetroleumProduct_Pipelines_US_EIA/", layer = "PetroleumProduct_Pipelines_US_202001")
-productsterminals <- rgdal::readOGR(dsn = "~/now/RTL/GIS_EIA/PetroleumProduct_Terminals_US_EIA/", layer = "PetroleumProduct_Terminals_US_202001")
-ngpipelines <- rgdal::readOGR(dsn = "~/now/RTL/GIS_EIA/NaturalGas_InterIntrastate_Pipelines_US_EIA/", layer = "NaturalGas_Pipelines_US_202001")
-ngstorage <- rgdal::readOGR(dsn = "~/now/RTL/GIS_EIA/NaturalGas_StorageRegions_US_EIA/", layer = "NaturalGas_StorageRegions_20151119")
-nghubs <- rgdal::readOGR(dsn = "~/now/RTL/GIS_EIA/NaturalGas_TradingHubs_US_EIA/", layer = "NaturalGas_TradingHubs_US_202002")
-lngterminals <- rgdal::readOGR(dsn = "~/now/RTL/GIS_EIA/Lng_ImportExportTerminals_US_EIA/", layer = "LNG_ImpExp_Terminals_US_202004")
-
-#usethis::use_data(abOilNG, overwrite = T)
-usethis::use_data(crudepipelines, overwrite = T)
-usethis::use_data(refineries, overwrite = T)
-usethis::use_data(productspipelines, overwrite = T)
-usethis::use_data(productsterminals, overwrite = T)
-usethis::use_data(ngpipelines, overwrite = T)
-usethis::use_data(ngstorage, overwrite = T)
-usethis::use_data(nghubs, overwrite = T)
-usethis::use_data(lngterminals, overwrite = T)
+# crudepipelines <- rgdal::readOGR(dsn = "~/now/RTL/GIS_EIA/CrudeOil_Pipelines_US_EIA/", layer = "CrudeOil_Pipelines_US_202001")
+# refineries <- rgdal::readOGR(dsn = "~/now/RTL/GIS_EIA/Petroleum_Refineries_US_EIA/", layer = "Petroleum_Refineries_US_2020")
+# productspipelines <- rgdal::readOGR(dsn = "~/now/RTL/GIS_EIA/PetroleumProduct_Pipelines_US_EIA/", layer = "PetroleumProduct_Pipelines_US_202001")
+# productsterminals <- rgdal::readOGR(dsn = "~/now/RTL/GIS_EIA/PetroleumProduct_Terminals_US_EIA/", layer = "PetroleumProduct_Terminals_US_202001")
+# ngpipelines <- rgdal::readOGR(dsn = "~/now/RTL/GIS_EIA/NaturalGas_InterIntrastate_Pipelines_US_EIA/", layer = "NaturalGas_Pipelines_US_202001")
+# ngstorage <- rgdal::readOGR(dsn = "~/now/RTL/GIS_EIA/NaturalGas_StorageRegions_US_EIA/", layer = "NaturalGas_StorageRegions_20151119")
+# nghubs <- rgdal::readOGR(dsn = "~/now/RTL/GIS_EIA/NaturalGas_TradingHubs_US_EIA/", layer = "NaturalGas_TradingHubs_US_202002")
+# lngterminals <- rgdal::readOGR(dsn = "~/now/RTL/GIS_EIA/Lng_ImportExportTerminals_US_EIA/", layer = "LNG_ImpExp_Terminals_US_202004")
+#
+# usethis::use_data(crudepipelines, overwrite = T)
+# usethis::use_data(refineries, overwrite = T)
+# usethis::use_data(productspipelines, overwrite = T)
+# usethis::use_data(productsterminals, overwrite = T)
+# usethis::use_data(ngpipelines, overwrite = T)
+# usethis::use_data(ngstorage, overwrite = T)
+# usethis::use_data(nghubs, overwrite = T)
+# usethis::use_data(lngterminals, overwrite = T)
 
 ## EIA Mapping
 tickers_eia <- read.csv('./data-raw/eia.csv',sep = ",",header = TRUE,na.strings = "NA",stringsAsFactors = FALSE)